+++ packages/components/src/components/sort-handle/sort-handle.browser.e2e.tsx
@@ -12,6 +12,7 @@ import {
   t9n,
   openClose,
   accessible,
+  topLayer,
 } from "../../tests/commonTests/browser";
 import { mockConsole } from "../../tests/utils/logging";
 import T9nStrings from "./assets/t9n/messages.en.json";
@@ -97,6 +98,10 @@ describe("openClose", () => {
   openClose((mountOptions) =>
     mount(<calcite-sort-handle label="test" set-position="4" set-size="10" />, mountOptions),
   );
+});
+
+describe("top layer placement", () => {
+  topLayer(() => mount(<calcite-sort-handle label="test" set-position="4" set-size="10" />));
 });
 
 describe("translation support", () => {