+++ packages/components-react/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.0-next.30](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.29...@esri/calcite-components-react@5.2.0-next.30) (2026-07-21)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.29](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.28...@esri/calcite-components-react@5.2.0-next.29) (2026-07-20)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.28](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.27...@esri/calcite-components-react@5.2.0-next.28) (2026-07-20)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.27](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.26...@esri/calcite-components-react@5.2.0-next.27) (2026-07-20)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.26](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.25...@esri/calcite-components-react@5.2.0-next.26) (2026-07-17)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.25](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.24...@esri/calcite-components-react@5.2.0-next.25) (2026-07-15)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.24](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.23...@esri/calcite-components-react@5.2.0-next.24) (2026-07-14)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.23](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.22...@esri/calcite-components-react@5.2.0-next.23) (2026-07-13)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.22](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.21...@esri/calcite-components-react@5.2.0-next.22) (2026-07-13)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.21](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.20...@esri/calcite-components-react@5.2.0-next.21) (2026-07-13)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.2.0-next.20](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.19...@esri/calcite-components-react@5.2.0-next.20) (2026-07-09)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [5.2.0-next.19](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.2.0-next.18...@esri/calcite-components-react@5.2.0-next.19) (2026-07-08)
 
 **Note:** Version bump only for package @esri/calcite-components-react